chore: use postbuild
